# openapi.yaml
openapi: 3.0.3
info:
  title: API Dígito Único
  description: |
    API REST para cálculo de dígito único com funcionalidades avançadas:

    ## Funcionalidades Principais
    - **Cálculo de Dígito Único**: Algoritmo otimizado para calcular o dígito único
    - **Gestão de Usuários**: CRUD completo de usuários
    - **Criptografia RSA**: Dados de usuário armazenados criptografados com chave pública RSA
    - **Cache Inteligente**: Cache LRU com capacidade de 10 entradas para otimizar cálculos
    - **Histórico**: Armazenamento de todos os cálculos realizados

    ## Algoritmo
    Dados `n` e `k`, forma-se `P` concatenando `n` exatamente `k` vezes e calcula-se o dígito único de `P`:
    - Se `P` tem um único dígito, ele é o resultado
    - Caso contrário, soma-se todos os dígitos e repete o processo

    **Exemplo:** n="9875", k=4
    - Soma dos dígitos de P: 9+8+7+5 × 4 = 116
    - digito_unico(116) → 1+1+6 = 8

    ## Fluxo de Criação de Usuário
    1. `POST /api/usuarios/chave-publica` — registra a chave pública RSA (cria o usuário, retorna o `id`)
    2. `POST /api/usuarios` — salva nome e email usando o `id` retornado (dados são criptografados)

    ## Autenticação
    Esta API não requer autenticação para fins de demonstração.

  version: 1.0.0
  contact:
    name: Equipe de Desenvolvimento
    email: dev@inter.com.br
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Servidor de Desenvolvimento

tags:
  - name: Usuários
    description: Operações de CRUD para usuários
  - name: Cálculos
    description: Operações de cálculo de dígito único
  - name: Criptografia
    description: Registro de chave pública RSA

paths:
  /api/usuarios/chave-publica:
    post:
      tags:
        - Criptografia
      summary: Registrar chave pública RSA
      description: |
        Cria um novo usuário registrando sua chave pública RSA.
        Retorna o `id` do usuário criado, necessário para o próximo passo (`POST /api/usuarios`).

        A chave pública deve estar em formato **Base64 puro (DER/X.509)**, sem cabeçalhos PEM.

        Geração via OpenSSL:
        ```
        openssl genrsa -out privada.pem 2048
        openssl rsa -in privada.pem -pubout -outform DER | base64 -w 0
        ```
      operationId: registrarChavePublica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChavePublicaRequest'
            example:
              chavePublica: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."
      responses:
        '201':
          description: Usuário criado com chave pública registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 1
                nomeCriptografado: null
                emailCriptografado: null
                chavePublica: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."
        '400':
          description: Chave pública inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/usuarios:
    get:
      tags:
        - Usuários
      summary: Listar todos os usuários
      description: Retorna uma lista com todos os usuários cadastrados no sistema
      operationId: listarUsuarios
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
              example:
                - id: 1
                  nomeCriptografado: "base64encodedEncryptedName=="
                  emailCriptografado: "base64encodedEncryptedEmail=="
                  chavePublica: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."

    post:
      tags:
        - Usuários
      summary: Salvar nome e email do usuário
      description: |
        Salva o nome e o email de um usuário já existente (criado via `POST /api/usuarios/chave-publica`).
        Os dados são criptografados com a chave pública RSA registrada anteriormente.

        **Pré-requisito:** o usuário deve ter sido criado via `POST /api/usuarios/chave-publica` e o `id` retornado deve ser informado no corpo da requisição.
      operationId: criarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequest'
            example:
              id: 1
              nome: "João Silva"
              email: "joao@example.com"
      responses:
        '201':
          description: Dados do usuário salvos e criptografados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
              example:
                id: 1
                nomeCriptografado: "base64encodedEncryptedName=="
                emailCriptografado: "base64encodedEncryptedEmail=="
                chavePublica: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."
        '400':
          description: Dados inválidos ou usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                message: "Dados de entrada inválidos"
                errors:
                  - field: "email"
                    message: "Email deve ter formato válido"
                  - field: "nome"
                    message: "Nome é obrigatório"

  /api/usuarios/{id}/descriptografar:
    post:
      tags:
        - Criptografia
      summary: Descriptografar dados do usuário
      description: |
        Retorna nome e email em texto claro usando a chave privada RSA do cliente.

        A chave privada deve estar em formato **Base64 puro (DER/PKCS8)**, sem cabeçalhos PEM.

        Extração via OpenSSL:
        ```
        openssl pkcs8 -topk8 -inform PEM -outform DER -in privada.pem -nocrypt | base64 -w 0
        ```
      operationId: descriptografarDados
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do usuário
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChavePrivadaRequest'
            example:
              chavePrivada: "MIIEvQIBADANBgkqhkiG9w0BAQEFAASC..."
      responses:
        '200':
          description: Dados descriptografados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioDTO'
              example:
                id: 1
                nome: "João Silva"
                email: "joao@example.com"
        '400':
          description: Chave privada inválida ou incompatível com a chave pública registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado

  /api/usuarios/{id}:
    get:
      tags:
        - Usuários
      summary: Buscar usuário por ID
      description: Retorna os dados de um usuário específico
      operationId: buscarUsuarioPorId
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do usuário
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Usuário não encontrado"
                timestamp: "2024-01-15T10:30:00Z"

    put:
      tags:
        - Usuários
      summary: Atualizar usuário
      description: |
        Atualiza nome e email de um usuário existente.
        Os novos dados são criptografados com a chave pública RSA do usuário.
      operationId: atualizarUsuario
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do usuário
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioAtualizarRequest'
            example:
              nome: "João Silva Atualizado"
              email: "joao.atualizado@example.com"
      responses:
        '200':
          description: Usuário atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Usuário não encontrado

    delete:
      tags:
        - Usuários
      summary: Deletar usuário
      description: Remove um usuário do sistema permanentemente
      operationId: deletarUsuario
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do usuário
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Usuário deletado com sucesso
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/calculos/digito-unico:
    post:
      tags:
        - Cálculos
      summary: Calcular dígito único
      description: |
        Calcula o dígito único de um número `n` concatenado `k` vezes.

        **Algoritmo:**
        1. Forma P = n concatenado k vezes
        2. Soma todos os dígitos de P
        3. Se o resultado tiver mais de um dígito, repete o processo
        4. Retorna o dígito único final (0-9)

        **Exemplo:** n="9875", k=4
        - Soma: (9+8+7+5) × 4 = 116
        - digito_unico(116) → 1+1+6 = 8

        **Parâmetros conforme especificação:**
        - `n`: string representando um inteiro (1 ≤ n ≤ 10^1000000)
        - `k`: número de concatenações (1 ≤ k ≤ 10^5)

        Resultados são armazenados em cache LRU (máximo 10 entradas). O campo `fromCache` indica se o resultado foi obtido do cache.
      operationId: calcularDigitoUnico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculoRequest'
            examples:
              exemplo_simples:
                summary: Exemplo simples
                value:
                  n: "123"
                  k: 2
              exemplo_com_resultado_8:
                summary: Exemplo da especificação (resultado esperado = 8)
                value:
                  n: "9875"
                  k: 4
      responses:
        '200':
          description: Cálculo realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculoResponse'
              example:
                n: "9875"
                k: 4
                digitoUnico: 8
                dataCalculo: "2024-01-15T10:30:00"
                fromCache: false
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                message: "Parâmetros inválidos"
                errors:
                  - field: "n"
                    message: "Parâmetro n deve conter apenas dígitos"
                  - field: "k"
                    message: "Parâmetro k deve ser maior que 0"

  /api/calculos/usuario/{usuarioId}:
    get:
      tags:
        - Cálculos
      summary: Buscar cálculos de um usuário
      description: Retorna o histórico de todos os cálculos realizados por um usuário específico
      operationId: buscarCalculosUsuario
      parameters:
        - name: usuarioId
          in: path
          required: true
          description: ID único do usuário
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Histórico de cálculos retornado com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resultado'
              example:
                - id: 1
                  n: "123"
                  k: 2
                  digitoUnico: 3
                  dataCalculo: "2024-01-15T10:30:00"
                - id: 2
                  n: "9875"
                  k: 4
                  digitoUnico: 8
                  dataCalculo: "2024-01-15T10:35:00"

components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do usuário
          example: 1
        nomeCriptografado:
          type: string
          nullable: true
          description: Nome do usuário criptografado com a chave pública RSA (Base64)
          example: "base64encodedEncryptedName=="
        emailCriptografado:
          type: string
          nullable: true
          description: Email do usuário criptografado com a chave pública RSA (Base64)
          example: "base64encodedEncryptedEmail=="
        chavePublica:
          type: string
          nullable: true
          description: Chave pública RSA do usuário em formato Base64 puro (DER/X.509)
          example: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."

    UsuarioRequest:
      type: object
      required:
        - id
        - nome
        - email
      properties:
        id:
          type: integer
          format: int64
          description: ID do usuário retornado em POST /api/usuarios/chave-publica
          example: 1
        nome:
          type: string
          description: Nome completo do usuário (será criptografado)
          minLength: 2
          maxLength: 100
          example: "João Silva"
        email:
          type: string
          format: email
          description: Email do usuário (será criptografado)
          example: "joao@example.com"

    UsuarioAtualizarRequest:
      type: object
      required:
        - nome
        - email
      properties:
        nome:
          type: string
          description: Novo nome do usuário (será criptografado)
          minLength: 2
          maxLength: 100
          example: "João Silva Atualizado"
        email:
          type: string
          format: email
          description: Novo email do usuário (será criptografado)
          example: "joao.atualizado@example.com"

    ChavePrivadaRequest:
      type: object
      required:
        - chavePrivada
      properties:
        chavePrivada:
          type: string
          description: Chave privada RSA em formato Base64 puro (DER/PKCS8), sem cabeçalhos PEM
          example: "MIIEvQIBADANBgkqhkiG9w0BAQEFAASC..."

    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do usuário
          example: 1
        nome:
          type: string
          description: Nome do usuário em texto claro (descriptografado)
          example: "João Silva"
        email:
          type: string
          description: Email do usuário em texto claro (descriptografado)
          example: "joao@example.com"

    ChavePublicaRequest:
      type: object
      required:
        - chavePublica
      properties:
        chavePublica:
          type: string
          description: Chave pública RSA em formato Base64 puro (DER/X.509), sem cabeçalhos PEM
          example: "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA..."

    CalculoRequest:
      type: object
      required:
        - n
        - k
      properties:
        n:
          type: string
          description: "Número para cálculo (apenas dígitos). Conforme especificação: 1 ≤ n ≤ 10^1000000"
          pattern: '^\d+$'
          minLength: 1
          example: "9875"
        k:
          type: integer
          description: "Número de concatenações. Conforme especificação: 1 ≤ k ≤ 10^5"
          minimum: 1
          maximum: 100000
          example: 4
        usuarioId:
          type: integer
          format: int64
          description: ID do usuário (opcional) — associa o cálculo ao histórico do usuário
          minimum: 1
          example: 1

    CalculoResponse:
      type: object
      properties:
        n:
          type: string
          description: Número original usado no cálculo
          example: "9875"
        k:
          type: integer
          description: Número de concatenações realizadas
          example: 4
        digitoUnico:
          type: integer
          description: Dígito único calculado (0-9)
          minimum: 0
          maximum: 9
          example: 8
        dataCalculo:
          type: string
          format: date-time
          description: Data e hora do cálculo
          example: "2024-01-15T10:30:00"
        fromCache:
          type: boolean
          description: Indica se o resultado foi obtido do cache LRU
          example: false

    Resultado:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do cálculo
          example: 1
        n:
          type: string
          description: Número usado no cálculo
          example: "9875"
        k:
          type: integer
          description: Número de concatenações
          example: 4
        digitoUnico:
          type: integer
          description: Dígito único calculado
          minimum: 0
          maximum: 9
          example: 8
        dataCalculo:
          type: string
          format: date-time
          description: Data e hora do cálculo
          example: "2024-01-15T10:30:00"

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensagem de erro
          example: "Recurso não encontrado"
        timestamp:
          type: string
          format: date-time
          description: Data e hora do erro
          example: "2024-01-15T10:30:00Z"
        details:
          type: string
          description: Detalhes adicionais do erro
          example: "Usuário com ID 999 não foi encontrado"

    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Mensagem geral de erro de validação
          example: "Dados de entrada inválidos"
        timestamp:
          type: string
          format: date-time
          description: Data e hora do erro
          example: "2024-01-15T10:30:00Z"
        errors:
          type: array
          description: Lista detalhada de erros de validação
          items:
            type: object
            properties:
              field:
                type: string
                description: Campo que causou o erro
                example: "email"
              message:
                type: string
                description: Mensagem específica do erro
                example: "Email deve ter formato válido"

security: []

externalDocs:
  description: Repositório no GitHub
  url: https://github.com/Sampaio-Thiago/digito-unico-api